"Fixed" angular momentum calculation
This is working as intended, but has actually degraded the experience as now the hands are too unstable and result in jittery behaviour when grabbing a (non-frozen) object. This should be fixed when A) rotational impulse limits are introduced and B) the hand joint applies an amount of its impulse to the held object.